feat: 업로드한 파일, 생성된 파일 화면 추가
